Testing of Player.java highlighted more bugs in Board.java, bugs fixed in both classes
minimax() can now be run, meaning the program can now play draughts!

The output has been used to test against an online opponent- it always makes legal moves, even though does not do very well.

Next step is to develop the graphics, with the aim of connecting the minimax output to icons on the board.