Added undo button!! Undoes the last human-done move only.
To make this work, the Game class had to be tweaked. When the game is in MultiJump mode, currentBoard is not updated as it needs to show the board state at the start of the turn, right until the end of the turn. Instead, temporayBoard and temporaryLegalMoves have been added, and if the game is in MultiJump then the getters return those instead of the originals. This minimises code changes outside of game.java.
To allow this, inMultiJump has been shifted from ActivityGame to Game.