Project write-up of development ongoing, Board class done and AI class nearly there.
Couple of changes to heuristic algorithm- variables for heuristic values have been moves outside the method, made static and final so that they do not have to be redeclared at every execution.
Also fixed a bug where when player 2 is winning the heuristic weights piece differences incorrectl;  documented in write-up.